- [ZBX-3477] fixed deletion of discovery rules without item prototypes
git-svn-id: svn://svn.zabbix.com/trunk@17283 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ frontends/php/api/classes/class.cdiscoveryrule.php
@@ -974,9 +974,10 @@ COpt::memoryPick();
 			while($item = DBfetch($db_items)){
 				$iprototypeids[$item['itemid']] = $item['itemid'];
 			}
-			if(!CItemPrototype::delete($iprototypeids, true))
-				self::exception(ZBX_API_ERROR_PARAMETERS, 'Cannot delete discovery rule');
-
+			if(!empty($iprototypeids)){
+				if(!CItemPrototype::delete($iprototypeids, true))
+					self::exception(ZBX_API_ERROR_PARAMETERS, 'Cannot delete discovery rule');
+			}
 
 			DB::delete('items', array('itemid'=>$ruleids));
